PASSPORT Authentication based on TOKENS (JWT)

+++ routes/leadersRouter.js
@@ -6,6 +6,7 @@ const Leaders = require('../models/leaders');
 
 const leadersRouter = express.Router();
 leadersRouter.use(bodyParser.json());
+
 leadersRouter.route('/')
     .get((req, res, next) => {
         Leaders.find({})
@@ -17,27 +18,37 @@ leadersRouter.route('/')
             .catch((err) => next(err));
     })
     .post(authenticate.verifyUser, (req, res, next) => {
-        Leaders.create(req.body)
-            .then((leader) => {
-                console.log('Leader created', leader);
-                res.statusCode = 200;
-                res.setHeader('Content-type', 'application/json');
-                res.json(leader);
-            }, (err) => next(err))
-            .catch((err) => next(err));
+        if (req.user.admin === true) {
+            Leaders.create(req.body)
+                .then((leader) => {
+                    console.log('Leader created', leader);
+                    res.statusCode = 200;
+                    res.setHeader('Content-type', 'application/json');
+                    res.json(leader);
+                }, (err) => next(err))
+                .catch((err) => next(err));
+        } else {
+            res.send('Only admins can adjust this page');
+        }
+
     })
     .put(authenticate.verifyUser, (req, res, next) => {
         res.statusCode = 403;
         res.end('Put operation does not make sense');
     })
     .delete(authenticate.verifyUser, (req, res, next) => {
-        Leaders.remove({})
-            .then((resp) => {
-                res.statusCode = 200;
-                res.setHeader('Content-type', 'application/json');
-                res.json(resp);
-            }, (err) => next(err))
-            .catch((err) => next(err));
+        if (req.user.admin === true) {
+            Leaders.remove({})
+                .then((resp) => {
+                    res.statusCode = 200;
+                    res.setHeader('Content-type', 'application/json');
+                    res.json(resp);
+                }, (err) => next(err))
+                .catch((err) => next(err));
+        } else {
+            res.send('Only admins can adjust this page');
+        }
+
     });
 
 
@@ -57,22 +68,31 @@ leadersRouter.route('/:id')                   // DAVOMIGA YOZING!
             + req.params.id);
     })
     .put(authenticate.verifyUser, (req, res, next) => {
-        Leaders.findByIdAndUpdate(req.params.id, { $set: req.body }, { new: true })
-            .then((leader) => {
-                res.statusCode = 200;
-                res.setHeader('Content-type', 'application/json');
-                res.json(leader);
-            }, (err) => next(err))
-            .catch((err) => next(err));
+        if (req.user.admin === true) {
+            Leaders.findByIdAndUpdate(req.params.id, { $set: req.body }, { new: true })
+                .then((leader) => {
+                    res.statusCode = 200;
+                    res.setHeader('Content-type', 'application/json');
+                    res.json(leader);
+                }, (err) => next(err))
+                .catch((err) => next(err));
+        } else {
+            res.send('Only admins can adjust this page');
+        }
     })
     .delete(authenticate.verifyUser, (req, res, next) => {
-        Leaders.findByIdAndRemove(req.params.id)
-            .then((resp) => {
-                res.statusCode = 200;
-                res.setHeader('Content-type', 'application/json');
-                res.json(resp);
-            }, (err) => next(err))
-            .catch((err) => next(err));
+        if (req.user.admin === true) {
+            Leaders.findByIdAndRemove(req.params.id)
+                .then((resp) => {
+                    res.statusCode = 200;
+                    res.setHeader('Content-type', 'application/json');
+                    res.json(resp);
+                }, (err) => next(err))
+                .catch((err) => next(err));
+        } else {
+            res.send('Only admins can adjust this page');
+        }
+
     });
 
 module.exports = leadersRouter;
