+++ routes/promoRouter.js
@@ -17,27 +17,35 @@ promotionRouter.route('/')
             .catch((err) => next(err));
     })
     .post(authenticate.verifyUser, (req, res, next) => {
-        Promotions.create(req.body)
-            .then((promotion) => {
-                console.log('Promotion created', promotion);
-                res.statusCode = 200;
-                res.setHeader('Content-type', 'application/json');
-                res.json(promotion);
-            }, (err) => next(err))
-            .catch((err) => next(err));
+        if (req.user.admin === true) {
+            Promotions.create(req.body)
+                .then((promotion) => {
+                    console.log('Promotion created', promotion);
+                    res.statusCode = 200;
+                    res.setHeader('Content-type', 'application/json');
+                    res.json(promotion);
+                }, (err) => next(err))
+                .catch((err) => next(err));
+        } else {
+            res.send('Only admins can adjust this page');
+        }
     })
     .put(authenticate.verifyUser, (req, res, next) => {
         res.statusCode = 403;
         res.end('Put operation does not make sense');
     })
     .delete(authenticate.verifyUser, (req, res, next) => {
-        Promotions.remove({})
-            .then((resp) => {
-                res.statusCode = 200;
-                res.setHeader('Content-type', 'application/json');
-                res.json(resp);
-            }, (err) => next(err))
-            .catch((err) => next(err));
+        if (req.user.admin === true) {
+            Promotions.remove({})
+                .then((resp) => {
+                    res.statusCode = 200;
+                    res.setHeader('Content-type', 'application/json');
+                    res.json(resp);
+                }, (err) => next(err))
+                .catch((err) => next(err));
+        } else {
+            res.send('Only admins can adjust this page');
+        }
     });
 
 
@@ -57,22 +65,32 @@ promotionRouter.route('/:id')                   // DAVOMIGA YOZING!
             + req.params.id);
     })
     .put(authenticate.verifyUser, (req, res, next) => {
-        Promotions.findByIdAndUpdate(req.params.id, { $set: req.body }, { new: true })
-            .then((promotion) => {
-                res.statusCode = 200;
-                res.setHeader('Content-type', 'application/json');
-                res.json(promotion);
-            }, (err) => next(err))
-            .catch((err) => next(err));
+        if (req.user.admin === true) {
+            Promotions.findByIdAndUpdate(req.params.id, { $set: req.body }, { new: true })
+                .then((promotion) => {
+                    res.statusCode = 200;
+                    res.setHeader('Content-type', 'application/json');
+                    res.json(promotion);
+                }, (err) => next(err))
+                .catch((err) => next(err));
+        } else {
+            res.send('Only admins can adjust this page');
+        }
+
     })
     .delete(authenticate.verifyUser, (req, res, next) => {
-        Promotions.findByIdAndRemove(req.params.id)
-            .then((resp) => {
-                res.statusCode = 200;
-                res.setHeader('Content-type', 'application/json');
-                res.json(resp);
-            }, (err) => next(err))
-            .catch((err) => next(err));
+        if (req.user.admin === true) {
+            Promotions.findByIdAndRemove(req.params.id)
+                .then((resp) => {
+                    res.statusCode = 200;
+                    res.setHeader('Content-type', 'application/json');
+                    res.json(resp);
+                }, (err) => next(err))
+                .catch((err) => next(err));
+        } else {
+            res.send('Only admins can adjust this page');
+        }
+
     });
